Refactor settings to be a promise we can wait on and fix tests for single-spa changes
