+++ packages/datagateway-dataview/src/state/actions/actions.test.tsx
@@ -15,30 +15,21 @@ import {
   ConfigurePluginHostSettingType,
   ConfigureFacilityImageSettingType,
 } from './actions.types';
-import axios from 'axios';
-import * as log from 'loglevel';
 import { actions, resetActions, dispatch, getState } from '../../setupTests';
-import {
-  loadUrls,
-  loadFacilityName,
-  MicroFrontendId,
-  RegisterRouteType,
-} from 'datagateway-common';
-import LogoLight from 'datagateway-common/src/images/datagateway-logo.svg';
-import LogoDark from 'datagateway-common/src/images/datgateway-white-text-blue-mark-logo.svg';
+import { loadUrls, loadFacilityName } from 'datagateway-common';
 
 jest.mock('loglevel');
 
-describe('Actions', () => {
-  beforeEach(() => {
-    global.document.dispatchEvent = jest.fn();
-    global.CustomEvent = jest.fn();
-  });
+const mockSettingsGetter = jest.fn();
+jest.mock('../../settings', () => ({
+  get settings() {
+    return mockSettingsGetter();
+  },
+}));
 
+describe('Actions', () => {
   afterEach(() => {
-    (axios.get as jest.Mock).mockReset();
-    (log.error as jest.Mock).mockReset();
-    (CustomEvent as jest.Mock).mockClear();
+    mockSettingsGetter.mockReset();
     resetActions();
   });
 
@@ -97,42 +88,29 @@ describe('Actions', () => {
     });
   });
 
-  it('settings are loaded and facilityName, loadFeatureSwitches, loadUrls, loadBreadcrumbSettings and settingsLoaded actions are sent', async () => {
-    (axios.get as jest.Mock)
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            facilityName: 'Generic',
-            facilityImageURL: 'test-image.jpg',
-            features: {},
-            idsUrl: 'ids',
-            apiUrl: 'api',
-            breadcrumbs: {
-              test: {
-                replaceEntityField: 'title',
-              },
-            },
-            downloadApiUrl: 'download-api',
-            selectAllSetting: false,
-            routes: [
-              {
-                section: 'section',
-                link: 'link',
-                displayName: 'displayName',
-              },
-            ],
-            pluginHost: 'http://localhost:3000/',
-          },
-        })
-      )
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            testSection: { test: 'string' },
-          },
-        })
-      );
-
+  it('settings are loaded and facilityName, loadFeatureSwitches, loadUrls, loadBreadcrumbSettings, loadSelectAllSetting and settingsLoaded actions are sent', async () => {
+    mockSettingsGetter.mockReturnValue({
+      facilityName: 'Generic',
+      facilityImageURL: 'test-image.jpg',
+      features: {},
+      idsUrl: 'ids',
+      apiUrl: 'api',
+      breadcrumbs: {
+        test: {
+          replaceEntityField: 'title',
+        },
+      },
+      downloadApiUrl: 'download-api',
+      selectAllSetting: false,
+      routes: [
+        {
+          section: 'section',
+          link: 'link',
+          displayName: 'displayName',
+        },
+      ],
+      pluginHost: 'http://localhost:3000/',
+    });
     const asyncAction = configureApp();
     await asyncAction(dispatch, getState);
 
@@ -160,280 +138,44 @@ describe('Actions', () => {
     expect(actions).toContainEqual(
       loadPluginHostSetting('http://localhost:3000/')
     );
-    expect(CustomEvent).toHaveBeenCalledTimes(1);
-    expect(CustomEvent).toHaveBeenLastCalledWith(MicroFrontendId, {
-      detail: {
-        type: RegisterRouteType,
-        payload: {
-          section: 'section',
-          link: 'link',
-          plugin: 'datagateway-dataview',
-          displayName: 'displayName',
-          order: 0,
-          helpSteps: [],
-          logoLightMode: 'http://localhost:3000/' + LogoLight,
-          logoDarkMode: 'http://localhost:3000/' + LogoDark,
-          logoAltText: 'DataGateway',
-        },
-      },
-    });
   });
 
-  it('settings are loaded despite no features', async () => {
-    (axios.get as jest.Mock)
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            facilityName: 'Generic',
-            idsUrl: 'ids',
-            apiUrl: 'api',
-            downloadApiUrl: 'download-api',
-            routes: [
-              {
-                section: 'section',
-                link: 'link',
-                displayName: 'displayName',
-              },
-            ],
-          },
-        })
-      )
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            testSection: { test: 'string' },
-          },
-        })
-      );
+  it("doesn't send loadSelectAllSetting, loadBreadcrumbSettings, loadPluginHostSetting, loadFacilityImageSetting and loadFeatureSwitches actions when they're not defined", async () => {
+    mockSettingsGetter.mockReturnValue({
+      facilityName: 'Generic',
+      idsUrl: 'ids',
+      apiUrl: 'api',
+      downloadApiUrl: 'download-api',
+    });
 
     const asyncAction = configureApp();
     await asyncAction(dispatch, getState);
 
     expect(actions.length).toEqual(3);
-    expect(actions).toContainEqual(loadFacilityName('Generic'));
-    expect(actions).toContainEqual(
-      loadUrls({
-        idsUrl: 'ids',
-        apiUrl: 'api',
-        downloadApiUrl: 'download-api',
-        icatUrl: '',
-      })
-    );
+    expect(
+      actions.every(({ type }) => type !== ConfigureSelectAllSettingType)
+    ).toBe(true);
+    expect(
+      actions.every(({ type }) => type !== ConfigureBreadcrumbSettingsType)
+    ).toBe(true);
+    expect(
+      actions.every(({ type }) => type !== ConfigureFeatureSwitchesType)
+    ).toBe(true);
+    expect(
+      actions.every(({ type }) => type !== ConfigurePluginHostSettingType)
+    ).toBe(true);
+    expect(
+      actions.every(({ type }) => type !== ConfigureFacilityImageSettingType)
+    ).toBe(true);
+
     expect(actions).toContainEqual(settingsLoaded());
   });
 
-  it('settings loaded and multiple routes registered with any helpSteps provided', async () => {
-    (axios.get as jest.Mock)
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            facilityName: 'Generic',
-            features: {},
-            idsUrl: 'ids',
-            apiUrl: 'api',
-            breadcrumbs: {
-              test: {
-                replaceEntityField: 'title',
-              },
-            },
-            downloadApiUrl: 'download-api',
-            selectAllSetting: false,
-            routes: [
-              {
-                section: 'section0',
-                link: 'link0',
-                displayName: 'displayName0',
-                order: 0,
-              },
-              {
-                section: 'section1',
-                link: 'link1',
-                displayName: 'displayName1',
-                order: 1,
-              },
-            ],
-            helpSteps: [{ target: '#id', content: 'content' }],
-          },
-        })
-      )
-      .mockImplementationOnce(() =>
-        Promise.resolve({
-          data: {
-            testSection: { test: 'string' },
-          },
-        })
-      );
-
+  it("doesn't send any actions when settings are undefined", async () => {
+    mockSettingsGetter.mockReturnValue(undefined);
     const asyncAction = configureApp();
     await asyncAction(dispatch, getState);
 
-    expect(CustomEvent).toHaveBeenCalledTimes(2);
-    expect(CustomEvent).toHaveBeenNthCalledWith(1, MicroFrontendId, {
-      detail: {
-        type: RegisterRouteType,
-        payload: {
-          section: 'section0',
-          link: 'link0',
-          plugin: 'datagateway-dataview',
-          displayName: 'displayName0',
-          order: 0,
-          helpSteps: [{ target: '#id', content: 'content' }],
-          logoLightMode: undefined,
-          logoDarkMode: undefined,
-          logoAltText: 'DataGateway',
-        },
-      },
-    });
-    expect(CustomEvent).toHaveBeenNthCalledWith(2, MicroFrontendId, {
-      detail: {
-        type: RegisterRouteType,
-        payload: {
-          section: 'section1',
-          link: 'link1',
-          plugin: 'datagateway-dataview',
-          displayName: 'displayName1',
-          order: 1,
-          helpSteps: [],
-          logoLightMode: undefined,
-          logoDarkMode: undefined,
-          logoAltText: 'DataGateway',
-        },
-      },
-    });
-  });
-
-  it('logs an error if facility name is not defined in settings.json and fails to be loaded', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() =>
-      Promise.resolve({
-        data: {},
-      })
-    );
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      'Error loading /datagateway-dataview-settings.json: facilityName is undefined in settings'
-    );
-  });
-
-  it('logs an error if urls are not defined in settings.json and fails to be loaded', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() =>
-      Promise.resolve({
-        data: {
-          facilityName: 'Generic',
-        },
-      })
-    );
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      'Error loading /datagateway-dataview-settings.json: One of the URL options (idsUrl, apiUrl, downloadApiUrl) is undefined in settings'
-    );
-  });
-
-  it('logs an error if no routes are defined in settings.json and fails to be loaded', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() =>
-      Promise.resolve({
-        data: {
-          facilityName: 'Generic',
-          idsUrl: 'ids',
-          apiUrl: 'api',
-          downloadApiUrl: 'download-api',
-        },
-      })
-    );
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      'Error loading /datagateway-dataview-settings.json: No routes provided in the settings'
-    );
-  });
-
-  it('logs an error if route has missing entries', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() =>
-      Promise.resolve({
-        data: {
-          facilityName: 'Generic',
-          idsUrl: 'ids',
-          apiUrl: 'api',
-          downloadApiUrl: 'download-api',
-          routes: [
-            {
-              section: 'section',
-              link: 'link',
-            },
-          ],
-        },
-      })
-    );
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      'Error loading /datagateway-dataview-settings.json: Route provided does not have all the required entries (section, link, displayName)'
-    );
-  });
-
-  it('logs an error if settings.json fails to be loaded', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() => Promise.reject({}));
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      expect.stringContaining(
-        `Error loading /datagateway-dataview-settings.json: `
-      )
-    );
-  });
-
-  it('logs an error if settings.json fails to be loaded with custom path', async () => {
-    process.env.REACT_APP_DATAVIEW_BUILD_DIRECTORY = '/custom/directory/';
-    (axios.get as jest.Mock).mockImplementationOnce(() => Promise.reject({}));
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      expect.stringContaining(
-        `Error loading /custom/directory/datagateway-dataview-settings.json: `
-      )
-    );
-    delete process.env.REACT_APP_DATAVIEW_BUILD_DIRECTORY;
-  });
-
-  it('logs an error if settings.json is invalid JSON object', async () => {
-    (axios.get as jest.Mock).mockImplementationOnce(() =>
-      Promise.resolve({
-        data: 1,
-      })
-    );
-
-    const asyncAction = configureApp();
-    await asyncAction(dispatch, getState);
-
-    expect(log.error).toHaveBeenCalled();
-    const mockLog = (log.error as jest.Mock).mock;
-    expect(mockLog.calls[0][0]).toEqual(
-      'Error loading /datagateway-dataview-settings.json: Invalid format'
-    );
+    expect(actions.length).toEqual(0);
   });
 });
