+++ packages/datagateway-dataview/src/state/actions/index.tsx
@@ -14,15 +14,8 @@ import {
   ConfigureFacilityImageSettingPayload,
   ConfigureFacilityImageSettingType,
 } from './actions.types';
-import {
-  loadUrls,
-  loadFacilityName,
-  MicroFrontendToken,
-} from 'datagateway-common';
+import { loadUrls, loadFacilityName } from 'datagateway-common';
 import { Action } from 'redux';
-import axios from 'axios';
-import * as log from 'loglevel';
-import jsrsasign from 'jsrsasign';
 import { settings } from '../../settings';
 
 export const settingsLoaded = (): Action => ({
@@ -74,113 +67,44 @@ export const loadFacilityImageSetting = (
   },
 });
 
-export const configureApp = (): ThunkResult<void> => {
-  return (dispatch) => {
-    // invalid settings.json
-    if (typeof settings !== 'object') {
-      throw Error('Invalid format');
-    }
+export const configureApp = (): ThunkResult<Promise<void>> => {
+  return async (dispatch) => {
+    const settingsResult = await settings;
+    if (settingsResult) {
+      dispatch(loadFacilityName(settingsResult['facilityName']));
 
-    // Get the facility name from settings.
-    if ('facilityName' in settings) {
-      dispatch(loadFacilityName(settings['facilityName']));
-    } else {
-      throw new Error('facilityName is undefined in settings');
-    }
+      // features is an optional setting
+      if (settingsResult?.['features'] !== undefined) {
+        dispatch(loadFeatureSwitches(settingsResult['features']));
+      }
 
-    // features is an optional setting
-    if ('features' in settings) {
-      dispatch(loadFeatureSwitches(settings['features']));
-    }
-
-    if (
-      'idsUrl' in settings &&
-      'apiUrl' in settings &&
-      'downloadApiUrl' in settings
-    ) {
       dispatch(
         loadUrls({
-          idsUrl: settings['idsUrl'],
-          apiUrl: settings['apiUrl'],
-          downloadApiUrl: settings['downloadApiUrl'],
+          idsUrl: settingsResult['idsUrl'],
+          apiUrl: settingsResult['apiUrl'],
+          downloadApiUrl: settingsResult['downloadApiUrl'],
           icatUrl: '', // we currently don't need icatUrl in dataview so just pass empty string for now
         })
       );
-    } else {
-      throw new Error(
-        'One of the URL options (idsUrl, apiUrl, downloadApiUrl) is undefined in settings'
-      );
+
+      // Dispatch the action to load the breadcrumb settings (optional settings).
+      if (settingsResult?.['breadcrumbs'] !== undefined) {
+        dispatch(loadBreadcrumbSettings(settingsResult['breadcrumbs']));
+      }
+
+      if (settingsResult?.['selectAllSetting'] !== undefined) {
+        dispatch(loadSelectAllSetting(settingsResult['selectAllSetting']));
+      }
+
+      if (settingsResult?.['pluginHost'] !== undefined) {
+        dispatch(loadPluginHostSetting(settingsResult['pluginHost']));
+      }
+
+      if (settingsResult?.['facilityImageURL'] !== undefined) {
+        dispatch(loadFacilityImageSetting(settingsResult['facilityImageURL']));
+      }
+
+      dispatch(settingsLoaded());
     }
-
-    // Dispatch the action to load the breadcrumb settings (optional settings).
-    if ('breadcrumbs' in settings) {
-      dispatch(loadBreadcrumbSettings(settings['breadcrumbs']));
-    }
-
-    if ('selectAllSetting' in settings) {
-      dispatch(loadSelectAllSetting(settings['selectAllSetting']));
-    }
-
-    if ('pluginHost' in settings) {
-      dispatch(loadPluginHostSetting(settings['pluginHost']));
-    }
-
-    if ('facilityImageURL' in settings) {
-      dispatch(loadFacilityImageSetting(settings['facilityImageURL']));
-    }
-
-    /* istanbul ignore if */
-    if (process.env.NODE_ENV === `development`) {
-      const splitUrl = settings.downloadApiUrl.split('/');
-      const icatUrl = `${splitUrl
-        .slice(0, splitUrl.length - 1)
-        .join('/')}/icat`;
-      axios
-        .post(
-          `${icatUrl}/session`,
-          `json=${JSON.stringify({
-            plugin: 'simple',
-            credentials: [{ username: 'root' }, { password: 'pw' }],
-          })}`,
-          {
-            headers: {
-              'Content-Type': 'application/x-www-form-urlencoded',
-            },
-          }
-        )
-        .then((response) => {
-          axios
-            .get(`${settings['apiUrl']}/sessions`, {
-              headers: {
-                Authorization: `Bearer ${response.data.sessionId}`,
-              },
-            })
-            .then(() => {
-              const jwtHeader = { alg: 'HS256', typ: 'JWT' };
-              const payload = {
-                sessionId: response.data.sessionId,
-                username: 'Thomas409',
-              };
-              const jwt = jsrsasign.KJUR.jws.JWS.sign(
-                'HS256',
-                jwtHeader,
-                payload,
-                'shh'
-              );
-
-              window.localStorage.setItem(MicroFrontendToken, jwt);
-            })
-            .catch((error) => {
-              log.error(
-                `datagateway-api cannot verify ICAT session id: ${error.message}.
-                     This is likely caused if datagateway-api is pointing to a
-                     different ICAT than the one used by the IDS/TopCAT`
-              );
-            });
-        })
-        .catch((error) => log.error(`Can't log in to ICAT: ${error.message}`));
-    }
-
-    dispatch(settingsLoaded());
   };
 };
