+++ packages/datagateway-download/src/App.tsx
@@ -35,7 +35,7 @@ class App extends Component<unknown, { hasError: boolean }> {
   }
 
   handler(e: Event): void {
-    // attempt to re-render the plugin if the corresponding div is present
+    // attempt to re-render the plugin if we get told to
     const action = (e as CustomEvent).detail;
     if (action.type === RequestPluginRerenderType) {
       // This is a temporary fix for the current issue with the tab indicator
