+++ packages/datagateway-download/src/index.tsx
@@ -5,11 +5,8 @@ import './i18n';
 import App from './App';
 import axios from 'axios';
 import jsrsasign from 'jsrsasign';
-
 import singleSpaReact from 'single-spa-react';
-
 import {
-  RequestPluginRerenderType,
   MicroFrontendId,
   MicroFrontendToken,
   PluginRoute,
@@ -45,25 +42,6 @@ const render = (): void => {
   }
 };
 
-// window.addEventListener('single-spa:routing-event', () => {
-//   // attempt to re-render the plugin if the corresponding div is present
-//   render();
-// });
-
-document.addEventListener(MicroFrontendId, (e) => {
-  // attempt to re-render the plugin if the corresponding div is present
-  const action = (e as CustomEvent).detail;
-  if (action.type === RequestPluginRerenderType) {
-    // This is a temporary fix for the current issue with the tab indicator
-    // not updating after the size of the page has been altered.
-    // This is issue is being tracked by material-ui (https://github.com/mui-org/material-ui/issues/9337).
-    setTimeout(() => {
-      window.dispatchEvent(new CustomEvent('resize'));
-    }, 125);
-    // render();
-  }
-});
-
 // Single-SPA bootstrap methods have no idea what type of inputs may be
 // pushed down from the parent app
 export function bootstrap(props: unknown): Promise<void> {
@@ -78,7 +56,8 @@ export function unmount(props: unknown): Promise<void> {
   return reactLifecycles.unmount(props);
 }
 
-const fetchSettings = (): Promise<void> => {
+// only export this for testing
+export const fetchSettings = (): Promise<DownloadSettings | void> => {
   const settingsPath = process.env.REACT_APP_DOWNLOAD_BUILD_DIRECTORY
     ? process.env.REACT_APP_DOWNLOAD_BUILD_DIRECTORY +
       'datagateway-download-settings.json'
@@ -179,14 +158,15 @@ const fetchSettings = (): Promise<void> => {
         );
       });
 
-      setSettings(settings);
+      return settings;
     })
     .catch((error) => {
       log.error(`Error loading ${settingsPath}: ${error.message}`);
     });
 };
 
-fetchSettings();
+const settings = fetchSettings();
+setSettings(settings);
 
 if (
   process.env.NODE_ENV === `development` ||
@@ -195,39 +175,44 @@ if (
   render();
 
   if (process.env.NODE_ENV === `development`) {
-    axios.get('./datagateway-download-settings.json').then((settings) => {
-      const icatUrl = `${settings.data.icatUrl}`;
-      axios
-        .post(
-          `${icatUrl}/session`,
-          `json=${JSON.stringify({
-            plugin: 'simple',
-            credentials: [{ username: 'root' }, { password: 'pw' }],
-          })}`,
-          {
-            headers: {
-              'Content-Type': 'application/x-www-form-urlencoded',
-            },
-          }
-        )
-        .then((response) => {
-          const jwtHeader = { alg: 'HS256', typ: 'JWT' };
-          const payload = {
-            sessionId: response.data.sessionId,
-            username: 'dev',
-          };
-          const jwt = jsrsasign.KJUR.jws.JWS.sign(
-            'HS256',
-            jwtHeader,
-            payload,
-            'shh'
-          );
+    settings.then((settingsResult) => {
+      if (settingsResult) {
+        const splitUrl = settingsResult.downloadApiUrl.split('/');
+        const icatUrl = `${splitUrl
+          .slice(0, splitUrl.length - 1)
+          .join('/')}/icat`;
+        axios
+          .post(
+            `${icatUrl}/session`,
+            `json=${JSON.stringify({
+              plugin: 'simple',
+              credentials: [{ username: 'root' }, { password: 'pw' }],
+            })}`,
+            {
+              headers: {
+                'Content-Type': 'application/x-www-form-urlencoded',
+              },
+            }
+          )
+          .then((response) => {
+            const jwtHeader = { alg: 'HS256', typ: 'JWT' };
+            const payload = {
+              sessionId: response.data.sessionId,
+              username: 'dev',
+            };
+            const jwt = jsrsasign.KJUR.jws.JWS.sign(
+              'HS256',
+              jwtHeader,
+              payload,
+              'shh'
+            );
 
-          window.localStorage.setItem(MicroFrontendToken, jwt);
-        })
-        .catch((error) =>
-          console.error(`Can't log in to ICAT: ${error.message}`)
-        );
+            window.localStorage.setItem(MicroFrontendToken, jwt);
+          })
+          .catch((error) =>
+            console.error(`Can't log in to ICAT: ${error.message}`)
+          );
+      }
     });
   }
 }
