+++ packages/datagateway-search/src/index.tsx
@@ -10,7 +10,7 @@ import App from './App';
 import * as log from 'loglevel';
 import singleSpaReact from 'single-spa-react';
 import axios from 'axios';
-import { setSettings } from './settings';
+import { SearchSettings, setSettings } from './settings';
 import {
   MicroFrontendId,
   PluginRoute,
@@ -45,37 +45,47 @@ const render = (): void => {
   }
 };
 
-// window.addEventListener('single-spa:routing-event', () => {
-//   // attempt to re-render the plugin if the route has changed
-//   console.log('single-spa:routing-event');
-//   render();
-// });
-
-// document.addEventListener('scigateway', (e) => {
-//   // attempt to re-render the plugin if scigateway tells us to
-//   const action = (e as CustomEvent).detail;
-//   if (action.type === 'scigateway:api:plugin_rerender') {
-//     console.log('scigateway:api:plugin_rerender');
-//     render();
-//   }
-// });
-
 // Single-SPA bootstrap methods have no idea what type of inputs may be
 // pushed down from the parent app
 export function bootstrap(props: unknown): Promise<void> {
   return reactLifecycles.bootstrap(props);
 }
 
-const fetchSettings = (): Promise<void> => {
-  console.log('fetching settings');
+// only export this for testing
+export const fetchSettings = (): Promise<SearchSettings | void> => {
   const settingsPath = process.env.REACT_APP_SEARCH_BUILD_DIRECTORY
     ? process.env.REACT_APP_SEARCH_BUILD_DIRECTORY +
       'datagateway-search-settings.json'
     : '/datagateway-search-settings.json';
   return axios
-    .get(settingsPath)
+    .get<SearchSettings>(settingsPath)
     .then((res) => {
       const settings = res.data;
+
+      // invalid settings.json
+      if (typeof settings !== 'object') {
+        throw Error('Invalid format');
+      }
+
+      // Ensure the facility name exists.
+      if (!('facilityName' in settings)) {
+        throw new Error('facilityName is undefined in settings');
+      }
+
+      // Ensure all API related URLs are present.
+      if (
+        !(
+          'idsUrl' in settings &&
+          'apiUrl' in settings &&
+          'downloadApiUrl' in settings &&
+          'icatUrl' in settings
+        )
+      ) {
+        throw new Error(
+          'One of the URL options (idsUrl, apiUrl, downloadApiUrl, icatUrl) is undefined in settings'
+        );
+      }
+
       if (Array.isArray(settings['routes']) && settings['routes'].length) {
         settings['routes'].forEach((route: PluginRoute, index: number) => {
           if ('section' in route && 'link' in route && 'displayName' in route) {
@@ -114,14 +124,15 @@ const fetchSettings = (): Promise<void> => {
       } else {
         throw new Error('No routes provided in the settings');
       }
-      setSettings(settings);
+      return settings;
     })
     .catch((error) => {
       log.error(`Error loading ${settingsPath}: ${error.message}`);
     });
 };
 
-fetchSettings();
+const settings = fetchSettings();
+setSettings(settings);
 
 export function mount(props: unknown): Promise<void> {
   return reactLifecycles.mount(props);
@@ -137,4 +148,58 @@ if (
 ) {
   render();
   log.setDefaultLevel(log.levels.DEBUG);
+
+  if (process.env.NODE_ENV === `development`) {
+    settings.then((settingsResult) => {
+      if (settingsResult) {
+        const apiUrl = settingsResult.apiUrl;
+        axios
+          .post(
+            `${settingsResult.icatUrl}/session`,
+            `json=${JSON.stringify({
+              plugin: 'simple',
+              credentials: [{ username: 'root' }, { password: 'pw' }],
+            })}`,
+            {
+              headers: {
+                'Content-Type': 'application/x-www-form-urlencoded',
+              },
+            }
+          )
+          .then((response) => {
+            axios
+              .get(`${apiUrl}/sessions`, {
+                headers: {
+                  Authorization: `Bearer ${response.data.sessionId}`,
+                },
+              })
+              .then(() => {
+                const jwtHeader = { alg: 'HS256', typ: 'JWT' };
+                const payload = {
+                  sessionId: response.data.sessionId,
+                  username: 'dev',
+                };
+                const jwt = jsrsasign.KJUR.jws.JWS.sign(
+                  'HS256',
+                  jwtHeader,
+                  payload,
+                  'shh'
+                );
+
+                window.localStorage.setItem('scigateway:token', jwt);
+              })
+              .catch((error) => {
+                log.error(
+                  `datagateway-api cannot verify ICAT session id: ${error.message}.
+               This is likely caused if datagateway-api is pointing to a
+               different ICAT than the one used by the IDS/TopCAT`
+                );
+              });
+          })
+          .catch((error) =>
+            log.error(`Can't log in to ICAT: ${error.message}`)
+          );
+      }
+    });
+  }
 }
