--- packages/datagateway-search/src/settings.js
@@ -1,4 +1,0 @@
-export let settings;
-export const setSettings = (newSettings) => {
-  settings = newSettings;
-};
