+++ packages/datagateway-search/src/state/actions/index.tsx
@@ -10,9 +10,6 @@ import {
 } from './actions.types';
 import { loadUrls, loadFacilityName } from 'datagateway-common';
 import { Action } from 'redux';
-import axios from 'axios';
-import * as log from 'loglevel';
-import jsrsasign from 'jsrsasign';
 import { settings } from '../../settings';
 
 export const settingsLoaded = (): Action => ({
@@ -46,99 +43,34 @@ export const loadMaxNumResults = (
   },
 });
 
-export const configureApp = (): ThunkResult<void> => {
-  return (dispatch) => {
-    // invalid settings.json
-    if (typeof settings !== 'object') {
-      throw Error('Invalid format');
-    }
+export const configureApp = (): ThunkResult<Promise<void>> => {
+  return async (dispatch) => {
+    const settingsResult = await settings;
+    if (settingsResult) {
+      dispatch(loadFacilityName(settingsResult['facilityName']));
 
-    // Get the facility name from settings.
-    if ('facilityName' in settings) {
-      dispatch(loadFacilityName(settings['facilityName']));
-    } else {
-      throw new Error('facilityName is undefined in settings');
-    }
-
-    if (
-      'idsUrl' in settings &&
-      'apiUrl' in settings &&
-      'downloadApiUrl' in settings
-    ) {
       dispatch(
         loadUrls({
-          idsUrl: settings['idsUrl'],
-          apiUrl: settings['apiUrl'],
-          downloadApiUrl: settings['downloadApiUrl'],
-          icatUrl: settings['icatUrl'],
+          idsUrl: settingsResult['idsUrl'],
+          apiUrl: settingsResult['apiUrl'],
+          downloadApiUrl: settingsResult['downloadApiUrl'],
+          icatUrl: settingsResult['icatUrl'],
         })
       );
-    } else {
-      throw new Error(
-        'One of the URL options (idsUrl, apiUrl, downloadApiUrl, icatUrl) is undefined in settings'
-      );
-    }
 
-    if ('selectAllSetting' in settings) {
-      dispatch(loadSelectAllSetting(settings['selectAllSetting']));
-    }
+      if (settingsResult?.['selectAllSetting'] !== undefined) {
+        dispatch(loadSelectAllSetting(settingsResult['selectAllSetting']));
+      }
 
-    if ('searchableEntities' in settings) {
-      dispatch(loadSearchableEntitites(settings['searchableEntities']));
-    }
+      if (settingsResult?.['searchableEntities'] !== undefined) {
+        dispatch(loadSearchableEntitites(settingsResult['searchableEntities']));
+      }
 
-    if ('maxNumResults' in settings) {
-      dispatch(loadMaxNumResults(settings['maxNumResults']));
-    }
+      if (settingsResult?.['maxNumResults'] !== undefined) {
+        dispatch(loadMaxNumResults(settingsResult['maxNumResults']));
+      }
 
-    /* istanbul ignore if */
-    if (process.env.NODE_ENV === `development`) {
-      const apiUrl = settings.apiUrl;
-      axios
-        .post(
-          `${settings.icatUrl}/session`,
-          `json=${JSON.stringify({
-            plugin: 'simple',
-            credentials: [{ username: 'root' }, { password: 'pw' }],
-          })}`,
-          {
-            headers: {
-              'Content-Type': 'application/x-www-form-urlencoded',
-            },
-          }
-        )
-        .then((response) => {
-          axios
-            .get(`${apiUrl}/sessions`, {
-              headers: {
-                Authorization: `Bearer ${response.data.sessionId}`,
-              },
-            })
-            .then(() => {
-              const jwtHeader = { alg: 'HS256', typ: 'JWT' };
-              const payload = {
-                sessionId: response.data.sessionId,
-                username: 'dev',
-              };
-              const jwt = jsrsasign.KJUR.jws.JWS.sign(
-                'HS256',
-                jwtHeader,
-                payload,
-                'shh'
-              );
-
-              window.localStorage.setItem('scigateway:token', jwt);
-            })
-            .catch((error) => {
-              log.error(
-                `datagateway-api cannot verify ICAT session id: ${error.message}.
-                   This is likely caused if datagateway-api is pointing to a
-                   different ICAT than the one used by the IDS/TopCAT`
-              );
-            });
-        })
-        .catch((error) => log.error(`Can't log in to ICAT: ${error.message}`));
+      dispatch(settingsLoaded());
     }
-    dispatch(settingsLoaded());
   };
 };
